+++ CHANGELOG.md
@@ -8,3 +8,4 @@
  - [feat] use phpstan for static analysis
  - [feat] documentation in readme file
  - [feat] throw exception whenever message level is not defined
+ - [feat] add row builder service